requirement.txt
psycopg2-binary
flask
openai
faker